Extract mock requests
