+++ data/test/github/index.test.js
@@ -1,16 +1,9 @@
-import dotEnv from 'dotenv';
-import nock from 'nock';
-import fs from 'fs';
-
+import { github } from '../mocks';
 import client from '../../github';
 
 describe('Github', () => {
   before(() => {
-    dotEnv.config();
-    const gitHubResponse = JSON.parse(fs.readFileSync('./test/fixtures/github.json', 'utf-8'));
-    nock('https://api.github.com/')
-      .get(`/users/gregstewart/events?access_token=${process.env.GITHUB_PERSONAL_TOKEN}`)
-      .reply(200, gitHubResponse);
+    github();
   });
 
   it('fetches my events', (done) => {
