+++ data/test/vimeo/index.test.js
@@ -1,17 +1,11 @@
-import dotEnv from 'dotenv';
-import nock from 'nock';
-import fs from 'fs';
+import { vimeo, invalidVimeo } from '../mocks';
 
 import client from '../../vimeo';
 
 describe('Vimeo', () => {
   describe('authorised request', () => {
     beforeEach(() => {
-      dotEnv.config();
-      const feed = fs.readFileSync('./test/fixtures/vimeo.json', 'utf-8');
-      nock('https://api.vimeo.com')
-        .get('/me/videos?page=1&per_page=10')
-        .reply(200, feed);
+      vimeo();
     });
 
     it('fetches my feed', (done) => {
@@ -30,14 +24,10 @@ describe('Vimeo', () => {
   describe('unauthorised request', () => {
     let invalidAuth;
     beforeEach(() => {
-      dotEnv.config();
       invalidAuth = {
         error: 'You must provide a valid authenticated access token.',
       };
-      process.env.VIMEO_ACCESS_TOKEN = 'foo';
-      nock('https://api.vimeo.com')
-        .get('/me/videos?page=1&per_page=10')
-        .reply(401, invalidAuth);
+      invalidVimeo(invalidAuth);
     });
     it('returns an invalid response', (done) => {
       client.get('/me/videos')
